Add HLO information gathering
Add the functions and information necessary to capture the HLO information
from the VistA instance.

To capture the action routine and tag information, an expansion was
necessary when attempting to capture information stored within a
"multiple" field within FileMan.

Add the capability to display the information within a "multiple" Fileman
field If it has a '/' between two numbers, '1/.04', ot will assume the 1
field is a multiple and attempt to pull out the .04 field information.

Change-Id: I470eaba94771b026f940d487238f2b160dc2b9ad