Add code to pull out patch dependency info
Add patch dependency information to the install_information.json file that
is used in the vivian display.  Ensure that the value and label fields are
kept to allow the install autocomplete to correctly display the values.

label is the name of the install
value is the index of the install in the package array.

Change-Id: If6d2b237aba23682ce32d9ac71d64a3d02151801