Fix date display in graphing and lab.
Both apparently started as a common code base. Both internally
implemented FMToDateTime when there is a correct version in ORFn.pas.

So I moved all references to FMToDateTime to use the versions in
ORFn.pas and deleted the original method.

There are some tiny changes to date format as c was too long to display
on graphs and worksheet cells. Changed c to ddddd hh:nn.